+++ src/translations/translations_3.rs
@@ -8,9 +8,13 @@ use crate::{Language, StyleType};
 // This is referred to settings (General settings)
 pub fn general_translation(language: Language) -> &'static str {
     match language {
-        Language::EN => "General",
+        Language::EN | Language::RO => "General",
         Language::ES => "Generales",
         Language::IT => "Generali",
+        Language::FR => "Général",
+        Language::DE => "Allgemein",
+        Language::PL => "Ogólne",
+        Language::RU => "Общие",
         Language::JA => "一般",
         _ => "General",
     }
@@ -18,7 +22,11 @@ pub fn general_translation(language: Language) -> &'static str {
 
 pub fn zoom_translation(language: Language) -> &'static str {
     match language {
-        Language::EN | Language::IT | Language::ES => "Zoom",
+        Language::EN | Language::IT | Language::ES | Language::FR | Language::DE | Language::RO => {
+            "Zoom"
+        }
+        Language::PL => "Powiększenie",
+        Language::RU => "Масштаб интерфейса",
         Language::JA => "ズーム",
         _ => "Zoom",
     }
@@ -29,6 +37,11 @@ pub fn mmdb_files_translation(language: Language) -> &'static str {
         Language::EN => "Database files",
         Language::ES => "Archivos de la base de datos",
         Language::IT => "File di database",
+        Language::FR => "Fichiers de la base de données",
+        Language::DE => "Datenbank Dateien",
+        Language::PL => "Pliki bazy danych",
+        Language::RU => "Файлы базы данных",
+        Language::RO => "Fișiere bază de date",
         Language::JA => "データベース ファイル",
         _ => "Database files",
     }
@@ -39,24 +52,26 @@ pub fn params_not_editable_translation(language: Language) -> &'static str {
         Language::EN => "The following parameters can't be modified during the analysis",
         Language::ES => "Los siguientes parámetros no pueden modificarse durante el análisis",
         Language::IT => "I seguenti parametri non sono modificabili durante l'analisi",
+        Language::FR => "Les paramètres suivants ne peuvent pas être modifiés durant l'analyse",
+        Language::DE => "Die folgenden Paramter können während der Analyse nicht verändert werden",
+        Language::PL => "Następujące parametry nie mogą być modyfikowane podczas analizy",
+        Language::RU => "Следующие параметры не могут быть изменены во время анализа трафика",
+        Language::RO => "Următorii parametri nu sunt modificabili în timpul analizei",
         Language::JA => "以下のパラメーターは分析中は変更できません",
         _ => "The following parameters can't be modified during the analysis",
     }
 }
-
-// pub fn file_path_translation(language: Language) -> &'static str {
-//     match language {
-//         Language::EN => "File path",
-//         Language::IT => "Percorso del file",
-//         _ => "File path",
-//     }
-// }
 
 pub fn custom_style_translation(language: Language) -> &'static str {
     match language {
         Language::EN => "Custom style",
         Language::ES => "Estilo personalizado",
         Language::IT => "Stile personalizzato",
+        Language::FR => "Style personnalisé",
+        Language::DE => "Benutzerdefinierter Stil",
+        Language::PL => "Niestandardowy styl",
+        Language::RU => "Свой стиль",
+        Language::RO => "Temă personalizată",
         Language::JA => "カスタム スタイル",
         _ => "Custom style",
     }
@@ -66,6 +81,10 @@ pub fn copy_translation(language: Language) -> &'static str {
     match language {
         Language::EN => "Copy",
         Language::IT | Language::ES => "Copia",
+        Language::FR | Language::RO => "Copie",
+        Language::DE => "Kopieren",
+        Language::PL => "Kopiuj",
+        Language::RU => "Скопировать",
         Language::JA => "コピー",
         _ => "Copy",
     }
@@ -73,9 +92,10 @@ pub fn copy_translation(language: Language) -> &'static str {
 
 pub fn port_translation(language: Language) -> &'static str {
     match language {
-        Language::EN => "Port",
+        Language::EN | Language::FR | Language::DE | Language::PL | Language::RO => "Port",
         Language::ES => "Puerto",
         Language::IT => "Porta",
+        Language::RU => "Порт",
         Language::JA => "ポート",
         _ => "Port",
     }
@@ -86,6 +106,11 @@ pub fn invalid_filters_translation(language: Language) -> &'static str {
         Language::EN => "Invalid filters",
         Language::ES => "Filtros inválidos",
         Language::IT => "Filtri non validi",
+        Language::FR => "Filtres invalides",
+        Language::DE => "Ungültige Filter",
+        Language::PL => "Nieprawidłowe filtry",
+        Language::RU => "Неверный формат фильтров",
+        Language::RO => "Filtre invalide",
         Language::JA => "無効なフィルター",
         _ => "Invalid filters",
     }
@@ -93,9 +118,13 @@ pub fn invalid_filters_translation(language: Language) -> &'static str {
 
 pub fn messages_translation(language: Language) -> &'static str {
     match language {
-        Language::EN => "Messages",
+        Language::EN | Language::FR => "Messages",
         Language::ES => "Mensajes",
         Language::IT => "Messaggi",
+        Language::DE => "Nachrichten",
+        Language::PL => "Wiadomości",
+        Language::RU => "Сообщения",
+        Language::RO => "Mesaje",
         Language::JA => "メッセージ",
         _ => "Messages",
     }
@@ -106,6 +135,11 @@ pub fn link_type_translation(language: Language) -> &'static str {
         Language::EN => "Link type",
         Language::ES => "Tipo de conexión",
         Language::IT => "Tipo di collegamento",
+        Language::FR => "Type de connexion",
+        Language::DE => "Verbindungsart",
+        Language::PL => "Rodzaj połączenia", // "Typ łącza"?
+        Language::RU => "Тип соединения",
+        Language::RO => "Tipul conexiunii",
         Language::JA => "リンク タイプ",
         _ => "Link type",
     }
@@ -119,6 +153,11 @@ pub fn unsupported_link_type_translation(
         Language::EN => "The link type associated with this adapter is not supported by Sniffnet yet...",
         Language::ES => "La conexión asociada con este adaptador aún no esta implementada en Sniffnet...",
         Language::IT => "Il tipo di collegamento associato a questo adattatore di rete non è ancora supportato da Sniffnet...",
+        Language::FR => "Le type de connexion associé à cet adaptateur n'est pas encore supporté par Sniffnet...",
+        Language::DE => "Die Verbindungsart dieses Adapters wird noch nicht von Sniffnet unterstützt",
+        Language::PL => "Rodzaj połączenia powiązany z tym adapterem nie jest jeszcze obsługiwany przez Sniffnet...",
+        Language::RU => "Тип соединения, связанный с этим адаптером, пока не поддерживается Sniffnet...",
+        Language::RO => "Tipul conexiunii asociate acestui adaptor de rețea nu este încă suportat de Sniffnet...",
         Language::JA => "このアダプターのリンク タイプは Sniffnet ではまだサポートされていません...",
         _ => "The link type associated with this adapter is not supported by Sniffnet yet...",
     }.to_string();
@@ -133,6 +172,11 @@ pub fn style_from_file_translation(language: Language) -> &'static str {
         Language::EN => "Select style from a file",
         Language::ES => "Selecciona el estilo desde un archivo",
         Language::IT => "Seleziona lo stile da un file",
+        Language::FR => "Sélectionner un style à partir d'un fichier",
+        Language::DE => "Stil aus einer Datei wählen",
+        Language::PL => "Wybierz styl z pliku",
+        Language::RU => "Выберите тему из файла",
+        Language::RO => "Selectează tema dintr-un fișier",
         Language::JA => "ファイルからスタイルを選択してください",
         _ => "Select style from a file",
     }
@@ -143,6 +187,11 @@ pub fn database_from_file_translation(language: Language) -> &'static str {
         Language::EN => "Select database file",
         Language::ES => "Selecciona un archivo de base de datos",
         Language::IT => "Seleziona file di database",
+        Language::FR => "Sélection d'un fichier de base de données",
+        Language::DE => "Datenbank Datei auswählen",
+        Language::PL => "Wybierz plik bazy danych",
+        Language::RU => "Выберите файл базы данных",
+        Language::RO => "Selectează fișier bază de date",
         Language::JA => "データベース ファイルを選択してください",
         _ => "Select database file",
     }
@@ -153,6 +202,11 @@ pub fn filter_by_host_translation(language: Language) -> &'static str {
         Language::EN => "Filter by network host",
         Language::ES => "Filtra por host de red",
         Language::IT => "Filtra per host di rete",
+        Language::FR => "Filtrer par réseau hôte",
+        Language::DE => "Nach Netzwerk-Host filtern",
+        Language::PL => "Filtruj według hosta sieciowego",
+        Language::RU => "Фильтр по сетевому хосту",
+        Language::RO => "Filtrează după host-ul de rețea",
         Language::JA => "ネットワーク ホストでフィルター",
         _ => "Filter by network host",
     }
@@ -160,9 +214,12 @@ pub fn filter_by_host_translation(language: Language) -> &'static str {
 
 pub fn service_translation(language: Language) -> &'static str {
     match language {
-        Language::EN => "Service",
+        Language::EN | Language::FR | Language::DE => "Service",
         Language::ES => "Servicio",
         Language::IT => "Servizio",
+        Language::PL => "Usługa",
+        Language::RU => "Сервис",
+        Language::RO => "Serviciu",
         Language::JA => "サービス",
         _ => "Service",
     }
@@ -172,6 +229,11 @@ pub fn export_capture_translation(language: Language) -> &'static str {
     match language {
         Language::EN => "Export capture file",
         Language::IT => "Esporta file di cattura",
+        Language::FR => "Exporter le fichier de capture",
+        Language::DE => "Aufzeichnungsdatei exportieren",
+        Language::PL => "Eksportuj plik przechwytywania",
+        Language::RU => "Экспорт файла захвата",
+        Language::RO => "Export fișier captură",
         Language::JA => "キャプチャ ファイルをエクスポート",
         _ => "Export capture file",
     }
@@ -182,6 +244,11 @@ pub fn directory_translation(language: Language) -> &'static str {
     match language {
         Language::EN => "Directory",
         Language::IT => "Cartella",
+        Language::FR => "Répertoire",
+        Language::DE => "Ordner",
+        Language::PL => "Katalog", // Katalog usually refers to Linux based directory while "Folder" is more common between Windows users (e.g. windows explorer refers to directories as "Folders")
+        Language::RU => "Директория",
+        Language::RO => "Director",
         Language::JA => "ディレクトリー",
         _ => "Directory",
     }
@@ -191,6 +258,11 @@ pub fn select_directory_translation(language: Language) -> &'static str {
     match language {
         Language::EN => "Select destination directory",
         Language::IT => "Seleziona cartella di destinazione",
+        Language::FR => "Sélectionner le répertoire de destination",
+        Language::DE => "Zielorder wählen",
+        Language::PL => "Wybierz katalog docelowy", // "Wybierz folder docelowy"?
+        Language::RU => "Выберите директорию назначения",
+        Language::RO => "Selectează directorul destinație",
         Language::JA => "宛先のディレクトリーを選択する",
         _ => "Select destination directory",
     }
@@ -200,6 +272,11 @@ pub fn file_name_translation(language: Language) -> &'static str {
     match language {
         Language::EN => "File name",
         Language::IT => "Nome del file",
+        Language::FR => "Nom du fichier",
+        Language::DE => "Dateiname",
+        Language::PL => "Nazwa pliku",
+        Language::RU => "Имя файла",
+        Language::RO => "Nume fișier",
         Language::JA => "ファイル ネーム",
         _ => "File name",
     }
@@ -209,6 +286,11 @@ pub fn thumbnail_mode_translation(language: Language) -> &'static str {
     match language {
         Language::EN => "Thumbnail mode",
         Language::IT => "Modalità miniatura",
+        Language::FR => "Mode miniature",
+        Language::DE => "Bild-in-Bild Modus",
+        Language::PL => "Tryb miniatury",
+        Language::RU => "Режим миниатюры",
+        Language::RO => "Mod thumbnail",
         Language::JA => "サムネイル モード",
         _ => "Thumbnail mode",
     }
@@ -218,6 +300,11 @@ pub fn learn_more_translation(language: Language) -> &'static str {
     match language {
         Language::EN => "Do you want to learn more?",
         Language::IT => "Vuoi saperne di più?",
+        Language::FR => "Voulez-vous en savoir davantage?",
+        Language::DE => "Mehr erfahren",
+        Language::PL => "Chcesz dowiedzieć się więcej?",
+        Language::RU => "Хотите узнать больше?",
+        Language::RO => "Vrei să înveți mai multe?",
         Language::JA => "もっと知りたいですか？",
         _ => "Do you want to learn more?",
     }